統一記憶體/字串操作介面並優化型別結構
本次提交：
- 移除多餘的標頭檔案，集中記憶體/字串操作於自定義函式（如 GUI__memcpy、GUI__strcpy 等），提升可移植性。
- 統一並優化結構體型別（如 int16_t、uintptr_t），修正 API 介面型別與命名一致性。
- 移除/簡化部分條件編譯，直接使用 WM 支援。
- 修正範例程式亂碼字串，提升語系相容性。
- 移除未使用檔案，調整專案檔案引用，提升可維護性。

+++ GUI/Widget/HEADER_Private.h
@@ -6,19 +6,21 @@
 #include "GUI_ARRAY.h"
 
 typedef struct {
-  int Width;
-  int16_t Align;
-  WM_HMEM hDrawObj;
-  char acText[1];
+	int16_t Width;
+	int16_t Align;
+	WM_HMEM hDrawObj;
+	char acText[1];
 } HEADER_COLUMN;
+
 typedef struct {
-  WIDGET Widget;
-  RGB_COLOR BkColor;
-  RGB_COLOR TextColor;
-  GUI_ARRAY Columns;
-  int CapturePosX;
-  int CaptureItem;
-  int ScrollPos;
-  const GUI_FONT  * pFont;
+	WIDGET Widget;
+	RGB_COLOR BkColor;
+	RGB_COLOR TextColor;
+	GUI_ARRAY Columns;
+	int16_t CapturePosX;
+	int16_t CaptureItem;
+	int16_t ScrollPos;
+	const GUI_FONT *pFont;
 } HEADER_Obj;
+
 void HEADER__SetDrawObj(HEADER_Handle hObj, unsigned Index, GUI_DRAW_HANDLE hDrawObj);
